pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = {extras = ["email"], version = "^2.12.4"}
pydantic-settings = "^2.11.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
fastapi = "^0.121.0"
asyncpg = "^0.30.0"
bcrypt = "4.0.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
